Running coverage tests separately from other tests
The coverage tests cannot run in the same testee process as the other
tests because the files get cached by the browser after the other tests
run and then never get instrumented for the coverage tests.
